realize funcion de la aplicacion y el apartado para ver los registros se puede borrar la cuenta desde el perfil asi tenga aprendices registrados

+++ php/delete_user.php
@@ -5,35 +5,59 @@ require('main.php');
 
 $user = limpiar_cadena($_POST['usuario_del']);
 $clave = limpiar_cadena($_POST['confirm_clave']);
+$delete = $_POST['delete'];
 
 $id = $_SESSION['usuario']['id_usuario'];
+$UsuarioLimpio = false;
 
-$sql_User = "SELECT * FROM usuarios WHERE id_usuario = $id ;";
-$check = mysqli_query($db, $sql_User);
-
-if($check && mysqli_num_rows($check)==1){
-    $datos_user = mysqli_fetch_assoc($check);
+if($delete == 'on'){
+    $sql_delete = "DELETE FROM aprendices WHERE id_usuario = $id;";
+    $execute = mysqli_query($db, $sql_delete);
+    if($delete){
+        $UsuarioLimpio = true;
+    }
 }
 
-$check_clave = password_verify($clave, $datos_user['clave_usuario']);
+if($UsuarioLimpio){
+
+    $sql_User = "SELECT * FROM usuarios WHERE id_usuario = $id ;";
+    $check = mysqli_query($db, $sql_User);
+
+    if($check && mysqli_num_rows($check)==1){
+        $datos_user = mysqli_fetch_assoc($check);
+    }
+
+    $check_clave = password_verify($clave, $datos_user['clave_usuario']);
 
 
-if($check_clave){
+    if($check_clave){
 
-    if($id == $datos_user['id_usuario']){
+        if($id == $datos_user['id_usuario']){
 
-        $sqlDel = "DELETE FROM usuarios WHERE id_usuario = $id AND usuario_usuario = '$user';";
-        $delete = mysqli_query($db, $sqlDel);
+            $sqlDel = "DELETE FROM usuarios WHERE id_usuario = $id AND usuario_usuario = '$user';";
+            $delete = mysqli_query($db, $sqlDel);
 
-        if($delete){
-            echo "
-            <script>
-            alert('El usuario se elimino con exito');
-            location.href = '../index.php?vista=login';
-            </script>
-            ";
-            
+            if($delete){
+                echo "
+                <script>
+                alert('El usuario se elimino con exito');
+                location.href = '../index.php?vista=login';
+                </script>
+                ";
+                
+            }
         }
+
+    }else{
+        $_SESSION['errorClave'] = "
+        <div class='message-header'>
+            <p>Error de autenticación</p>
+        </div>
+        <div class='message-body'>
+            La <strong>CONTRASEÑA </strong> no es la correcta!
+        </div>
+        ";
+        header('location:../index.php?vista=perfil');
     }
 
 }else{
@@ -42,7 +66,7 @@ if($check_clave){
         <p>Error de autenticación</p>
     </div>
     <div class='message-body'>
-        La <strong>CONTRASEÑA </strong> no es la correcta!
+        El <strong>Usuario </strong> no se puede borrar porque tiene aprendices registrados!
     </div>
     ";
     header('location:../index.php?vista=perfil');
